requirements.txt
agora-token-builder==1.0.0
asgiref==3.4.1
Django==4.0.1
sqlparse==0.4.2
tzdata==2021.5
gtts
playsound
wikipedia
requests
bs4
googletrans
bing_image_downloader
pillow
googlesearch-python
SpeechRecognition
boto3==1.24.56
folium==0.12.1.post1
geojson==2.5.0
numpy==1.23.2
pandas==1.4.3
pyproj==3.3.1
pystac==1.6.1
rasterio==1.2.10
Shapely==1.8.4
streamlit==1.11.1
streamlit-folium==0.6.14
